+++ src/main/java/kpl/fiml/payment/application/PaymentService.java
@@ -3,8 +3,8 @@ package kpl.fiml.payment.application;
 import kpl.fiml.payment.domain.Payment;
 import kpl.fiml.payment.domain.PaymentRepository;
 import kpl.fiml.payment.domain.PaymentStatus;
-import kpl.fiml.payment.dto.PaymentCreateRequest;
-import kpl.fiml.payment.dto.PaymentDto;
+import kpl.fiml.payment.dto.request.PaymentCreateRequest;
+import kpl.fiml.payment.dto.response.PaymentDto;
 import kpl.fiml.sponsor.application.SponsorService;
 import kpl.fiml.sponsor.domain.Sponsor;
 import kpl.fiml.user.application.UserService;
