+++ src/main/java/kpl/fiml/payment/presentation/PaymentController.java
@@ -1,7 +1,7 @@
 package kpl.fiml.payment.presentation;
 
 import kpl.fiml.payment.application.PaymentService;
-import kpl.fiml.payment.dto.PaymentDto;
+import kpl.fiml.payment.dto.response.PaymentDto;
 import kpl.fiml.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
