+++ src/main/java/kpl/fiml/settlement/presentation/SettlementController.java
@@ -1,7 +1,7 @@
 package kpl.fiml.settlement.presentation;
 
 import kpl.fiml.settlement.application.SettlementService;
-import kpl.fiml.settlement.dto.SettlementDto;
+import kpl.fiml.settlement.dto.response.SettlementDto;
 import kpl.fiml.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
